Improve stcToBuffer() and stcToChars() to handle last one space
And move them into hdrgen.c, that's more better location.
