+++ src/attrib.c
@@ -394,95 +394,6 @@ Scope *StorageClassDeclaration::newScope(Scope *sc)
     return createNewScope(sc, scstc, sc->linkage, sc->protection, sc->explicitProtection, sc->structalign, sc->inlining);
 }
 
-/*************************************************
- * Pick off one of the storage classes from stc,
- * and return a pointer to a string representation of it.
- * stc is reduced by the one picked.
- * tmp[] is a buffer big enough to hold that string.
- */
-const char *StorageClassDeclaration::stcToChars(char tmp[], StorageClass& stc)
-{
-    struct SCstring
-    {
-        StorageClass stc;
-        TOK tok;
-        const char *id;
-    };
-
-    static SCstring table[] =
-    {
-        { STCauto,         TOKauto },
-        { STCscope,        TOKscope },
-        { STCstatic,       TOKstatic },
-        { STCextern,       TOKextern },
-        { STCconst,        TOKconst },
-        { STCfinal,        TOKfinal },
-        { STCabstract,     TOKabstract },
-        { STCsynchronized, TOKsynchronized },
-        { STCdeprecated,   TOKdeprecated },
-        { STCoverride,     TOKoverride },
-        { STClazy,         TOKlazy },
-        { STCalias,        TOKalias },
-        { STCout,          TOKout },
-        { STCin,           TOKin },
-        { STCmanifest,     TOKenum },
-        { STCimmutable,    TOKimmutable },
-        { STCshared,       TOKshared },
-        { STCnothrow,      TOKnothrow },
-        { STCwild,         TOKwild },
-        { STCpure,         TOKpure },
-        { STCref,          TOKref },
-        { STCtls },
-        { STCgshared,      TOKgshared },
-        { STCnogc,         TOKat,       "nogc" },
-        { STCproperty,     TOKat,       "property" },
-        { STCsafe,         TOKat,       "safe" },
-        { STCtrusted,      TOKat,       "trusted" },
-        { STCsystem,       TOKat,       "system" },
-        { STCdisable,      TOKat,       "disable" },
-        { 0,               TOKreserved }
-    };
-
-    for (int i = 0; table[i].stc; i++)
-    {
-        StorageClass tbl = table[i].stc;
-        assert(tbl & STCStorageClass);
-        if (stc & tbl)
-        {
-            stc &= ~tbl;
-            if (tbl == STCtls)  // TOKtls was removed
-                return "__thread";
-
-            TOK tok = table[i].tok;
-            if (tok == TOKat)
-            {
-                tmp[0] = '@';
-                strcpy(tmp + 1, table[i].id);
-                return tmp;
-            }
-            else
-                return Token::toChars(tok);
-        }
-    }
-    //printf("stc = %llx\n", (unsigned long long)stc);
-    return NULL;
-}
-
-void StorageClassDeclaration::stcToCBuffer(OutBuffer *buf, StorageClass stc)
-{
-    while (stc)
-    {
-        const size_t BUFFER_LEN = 20;
-        char tmp[BUFFER_LEN];
-        const char *p = stcToChars(tmp, stc);
-        if (!p)
-            break;
-        assert(strlen(p) < BUFFER_LEN);
-        buf->writestring(p);
-        buf->writeByte(' ');
-    }
-}
-
 /********************************* DeprecatedDeclaration ****************************/
 
 DeprecatedDeclaration::DeprecatedDeclaration(Expression *msg, Dsymbols *decl)
