+++ src/attrib.h
@@ -68,8 +68,6 @@ public:
     Scope *newScope(Scope *sc);
     bool oneMember(Dsymbol **ps, Identifier *ident);
 
-    static const char *stcToChars(char tmp[], StorageClass& stc);
-    static void stcToCBuffer(OutBuffer *buf, StorageClass stc);
     void accept(Visitor *v) { v->visit(this); }
 };
