+++ src/import.c
@@ -22,6 +22,7 @@
 #include "declaration.h"
 #include "id.h"
 #include "attrib.h"
+#include "hdrgen.h"
 
 /********************************* Import ****************************/
 
@@ -323,7 +324,10 @@ void Import::semantic(Scope *sc)
         protectionToBuffer(ob, Prot(protection));
         ob->writeByte(' ');
         if (isstatic)
-            StorageClassDeclaration::stcToCBuffer(ob, STCstatic);
+        {
+            stcToBuffer(ob, STCstatic);
+            ob->writeByte(' ');
+        }
         ob->writestring(": ");
 
         if (packages)
